split servo controls into its own component
